+++ libs/ngx-editorjs-demo/src/lib/config/index.ts
@@ -1,5 +1,4 @@
-import { PluginConfig } from '@tinynodes/ngx-editorjs';
-import { PluginCode, PluginLink, PluginMarker, PluginSimpleImage } from '@tinynodes/ngx-editorjs-plugins';
+import { PluginConfig, PluginCode, PluginLink, PluginMarker, PluginSimpleImage } from '@tinynodes/ngx-editorjs-plugins';
 
 /**
  * Creates the tools for the `ngx-tinynodes` demo application