+++ libs/ngx-editorjs-demo/src/lib/ngx-editorjs-demo.module.ts
@@ -3,12 +3,13 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule, MatCardModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
-import { NgxEditorJSModule, UserPlugins } from '@tinynodes/ngx-editorjs';
+import { NgxEditorJSModule } from '@tinynodes/ngx-editorjs';
 import {
   PluginCodeModule,
   PluginLinkModule,
   PluginMarkerModule,
-  PluginSimpleImageModule
+  PluginSimpleImageModule,
+  UserPlugins
 } from '@tinynodes/ngx-editorjs-plugins';
 import { MarkdownModule } from 'ngx-markdown';
 import { createNgxEditorJSDemoTools } from './config';