+++ libs/ngx-editorjs-plugins/src/lib/ngx-editorjs-plugins.module.ts
@@ -8,6 +8,7 @@ import { PluginListModule } from './plugins/list/list.module';
 import { PluginMarkerModule } from './plugins/marker/marker.module';
 import { PluginParagraphModule } from './plugins/paragraph/paragraph.module';
 import { PluginSimpleImageModule } from './plugins/simple-image/simple-image.module';
+import { NgxEditorJSPluginServiceModule } from './services/plugins.module';
 
 /**
  * The `NgxEditorjsPluginsModule` provides several additional modules
@@ -26,7 +27,8 @@ import { PluginSimpleImageModule } from './plugins/simple-image/simple-image.mod
     PluginListModule,
     PluginMarkerModule,
     PluginParagraphModule,
-    PluginSimpleImageModule
+    PluginSimpleImageModule,
+    NgxEditorJSPluginServiceModule
   ],
   exports: [
     PluginCodeModule,
@@ -36,7 +38,8 @@ import { PluginSimpleImageModule } from './plugins/simple-image/simple-image.mod
     PluginListModule,
     PluginMarkerModule,
     PluginParagraphModule,
-    PluginSimpleImageModule
+    PluginSimpleImageModule,
+    NgxEditorJSPluginServiceModule
   ]
 })
 export class NgxEditorjsPluginsModule {}