+++ libs/ngx-editorjs-plugins/src/lib/plugins/image/image.plugin.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ToolSettings } from '@editorjs/editorjs';
 import Image from '@editorjs/image';
-import { BasePlugin } from '@tinynodes/ngx-editorjs';
+import { BasePlugin } from '../../types/plugins';
 
 /**
  * This plugin returns the tool settings for a `<h1> - <h6>` block tool for `EditorJS`