+++ libs/ngx-editorjs-plugins/src/lib/plugins/simple-image/simple-image.plugin.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ToolSettings } from '@editorjs/editorjs';
 import SimpleImage from '@editorjs/simple-image';
-import { BasePlugin } from '@tinynodes/ngx-editorjs';
+import { BasePlugin } from '../../types/plugins';
 
 /**
  * This plugin returns the simple image tool for `<img>` block editor `EditorJS`