+++ libs/ngx-editorjs-plugins/src/lib/services/plugins.module.ts
@@ -1,0 +1,7 @@
+import { NgModule } from '@angular/core';
+import { NgxEditorJSPluginService } from './plugins.service';
+
+@NgModule({
+  providers: [NgxEditorJSPluginService]
+})
+export class NgxEditorJSPluginServiceModule {}