+++ libs/ngx-editorjs/src/index.ts
@@ -2,7 +2,5 @@ export { NgxEditorJSComponent } from './lib/containers/editorjs-component/editor
 export { NgxEditorJSDirective } from './lib/directives/ngx-editorjs.directive';
 export { NgxEditorJSModule } from './lib/ngx-editorjs.module';
 export { NgxEditorJSService } from './lib/services/editorjs.service';
-export { PluginService as ɵPluginService } from './lib/services/plugins.service';
 export { Block, HeaderBlock, ImageBlock, ListBlock, ListStyle, ParagraphBlock } from './lib/types/blocks';
 export { EditorJSConfig, NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './lib/types/config';
-export { BasePlugin, PluginConfig, UserPlugins } from './lib/types/plugins';