+++ libs/ngx-editorjs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed default plugins from project. All plugins are now provided by the `@tinynodes/ngx-editorjs-plugins` module
+- Removed `PluginService` from core and moved to `@tinynodes/ngx-editorjs-plugins`
 
 ## [1.1.2] - 2019-05-11